requirements.txt
absl-py==1.1.0
aiohttp==3.7.4.post0
alabaster==0.7.12
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
bagpy==0.4.10
beautifulsoup4==4.11.1
bitstring==3.1.9
bleach==5.0.1
blinker==1.4
blobfile==1.3.3
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.0.0
catkin-pkg==0.5.2
certifi==2022.12.7
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==1.6.0
commentjson==0.9.0
commonmark==0.9.1
ConfigArgParse==1.5.3
cryptography==36.0.0
cycler==0.11.0
cytoolz==0.11.0
DALL-E==0.1
dask==2.30.0
dearpygui==1.4.0
debugpy==1.6.1
decorator==4.4.2
defusedxml==0.7.1
distro==1.7.0
docutils==0.17.1
einops==0.4.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.8.0
fonttools==4.30.0
future==0.18.2
ghp-import==2.1.0
gnupg==2.3.1
google-auth==2.7.0
google-auth-oauthlib==0.4.6
grpcio==1.42.0
guppy3==3.1.2
h5py==3.7.0
hdf5plugin==3.3.1
idna==3.3
imagecodecs==2021.8.26
imageio==2.16.1
imageio-ffmpeg==0.4.7
imagesize==1.3.0
importlib-metadata==4.11.4
iniconfig==1.1.1
ipykernel==6.15.0
ipympl==0.9.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.6.2
jupyter-client==7.3.4
jupyter-core==4.11.0
jupyter-server==1.18.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
kiwisolver==1.3.2
kornia==0.6.4
lark-parser==0.7.8
llvmlite==0.38.1
loguru==0.5.3
lpips==0.1.4
lz4==4.0.1
m2r2==0.3.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mergedeep==1.3.4
mistune==0.8.4
mkdocs==1.3.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==5.2.0
mypy==0.971
mypy-extensions==0.4.3
natsort==8.1.0
nbclassic==0.4.2
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.5.1
ninja==1.10.2.3
notebook==6.4.12
notebook-shim==0.1.0
numba==0.55.2
numpy==1.21.2
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==22.3.1
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py3rosmsgs==1.18.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pycryptodomex==3.14.1
pydash==5.1.0
pyglet==1.5.16
Pygments==2.11.2
PyJWT==2.4.0
PyMCubes==0.1.2
pymdown-extensions==9.5
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
pysdf==0.1.8
pyserial==3.5
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
pytz==2021.3
pyu2f==0.1.5
pyvista==0.33.3
PyWavelets==1.3.0
PyYAML==5.4.1
pyyaml_env_tag==0.1
pyzmq==23.2.0
recommonmark==0.7.1
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.0.0
rinoh-typeface-dejavuserif==0.1.3
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.5.3
rospkg==1.4.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.7.3
scooby==0.5.12
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==58.0.4
Shapely==1.7.1
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==3.2.1
sphinx-autodoc-typehints==1.4.0
sphinx-bootstrap-theme==0.8.1
sphinx-markdown-parser==0.2.4
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.2.0
tensorboard==2.9.1
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2020.10.1
timm==0.6.7
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.1
torch==1.11.0
torch-ema==0.3
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.2
tqdm==4.63.0
traitlets==5.2.2.post1
trimesh==3.10.2
typing-extensions==3.10.0.2
unify==0.5
untokenize==0.1.1
urllib3==1.26.8
vtk==9.0.3
watchdog==2.1.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
wheel==0.37.1
widgetsnbextension==3.6.1
xmltodict==0.12.0
yapf==0.32.0
yarl==1.6.3
ytsphinx==1.2.1.dev20200430
zipp==3.8.0
zstandard==0.17.0
